CSL now uses zlib to help it compress image files, and both saved images and FASL files are packed using the code in serialize.cpp. As compared to the situation a few weeks ago this often causes image files to be smaller by about a factor of 2! The way in which the word-length and byte-order of the computer used to reload an image file is also a lot cleaner. This has involved changes in quite a few places within CSL and there will still be some improvements that I can make over the coming weeks, but I hope that this is the major step and future changes will count as more localized refinements.
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@3822 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/md5.cpp
@@ -35,7 +35,8 @@
 // compile-time configuration.
 //
 
-// passed through astyle and otherwise had formatting changes.
+// passed through astyle and otherwise had formatting changes. use size_t for
+// lengths.
 // A C Norman, 2015
 
 #ifndef HAVE_OPENSSL
@@ -93,7 +94,7 @@
 // This processes one or more 64-byte data blocks, but does NOT update
 // the bit counters.  There are no alignment requirements.
 //
-static const void *body(MD5_CTX *ctx, const void *data, unsigned long size)
+static const void *body(MD5_CTX *ctx, const void *data, size_t size)
 {   const unsigned char *ptr;
     MD5_u32plus a, b, c, d;
     MD5_u32plus saved_a, saved_b, saved_c, saved_d;
@@ -210,9 +211,9 @@ void MD5_Init(MD5_CTX *ctx)
     ctx->hi = 0;
 }
 
-void MD5_Update(MD5_CTX *ctx, const void *data, unsigned long size)
+void MD5_Update(MD5_CTX *ctx, const void *data, size_t size)
 {   MD5_u32plus saved_lo;
-    unsigned long used, available;
+    size_t used, available;
 
     saved_lo = ctx->lo;
     if ((ctx->lo = (saved_lo + size) & 0x1fffffff) < saved_lo)
@@ -244,7 +245,7 @@ void MD5_Update(MD5_CTX *ctx, const void *data, unsigned long size)
 }
 
 void MD5_Final(unsigned char *result, MD5_CTX *ctx)
-{   unsigned long used, available;
+{   size_t used, available;
 
     used = ctx->lo & 0x3f;
